Poistettu kirjan ja artikkelin toisteista koodia
