+++ src/main/java/lukuvinkkiKirjasto/domain/Artikkeli.java
@@ -1,63 +1,21 @@
 package lukuvinkkiKirjasto.domain;
 
 import java.time.LocalDate;
-import java.util.Date;
 
-public class Artikkeli {
+public class Artikkeli extends Tuote {
 
     private Integer id;
-    private String nimi;
-    private int pituus;
-    private String linkki;
-    private String tekija;
     private String julkaisuLehti;
-    private int julkaisuVuosi;
     private int numero;
     private String sivut;
-    private LocalDate paivamaara;
-    private boolean luettu;
-    private String luettuAika;
 
-    public Artikkeli(Integer id, String nimi, int pituus, String linkki, String 
-            tekija, String julkaisuLehti, int julkaisuVuosi, int numero, String 
-            sivut, LocalDate paivamaara, boolean luettu, String aika) {
-        
+    public Artikkeli(Integer id, String nimi, int pituus, String linkki, String tekija, String julkaisuLehti, int julkaisuVuosi, int numero, String sivut, LocalDate paivamaara, boolean luettu, String luettuAika) {
+
+        super(nimi, pituus, linkki, tekija, julkaisuVuosi, paivamaara, luettu, luettuAika);
         this.id = id;
-        this.nimi = nimi;
-        this.pituus = pituus;
-        this.linkki = linkki;
-        this.tekija = tekija;
         this.julkaisuLehti = julkaisuLehti;
-        this.julkaisuVuosi = julkaisuVuosi;
         this.numero = numero;
         this.sivut = sivut;
-        this.paivamaara = paivamaara;
-        this.luettu = luettu;
-        this.luettuAika = aika;
-    }
-
-    public boolean isLuettu() {
-        return luettu;
-    }
-
-    public void setLuettu(boolean luettu) {
-        this.luettu = luettu;
-    }
-    
-    public String getLuettuString() {
-        if (this.luettu) {
-            return "Luettu " + luettuAika;
-        } else {
-            return "Lukematon";
-        }
-    }
-
-    public String getLuettuAika() {
-        return luettuAika;
-    }
-
-    public void setLuettuAika(String luettuAika) {
-        this.luettuAika = luettuAika;
     }
 
     public int getNumero() {
@@ -84,14 +42,6 @@ public class Artikkeli {
         this.sivut = sivut;
     }
 
-    public int getJulkaistu() {
-        return julkaisuVuosi;
-    }
-
-    public void setJulkaistu(int julkaistu) {
-        this.julkaisuVuosi = julkaistu;
-    }
-
     public Integer getId() {
         return id;
     }
@@ -100,52 +50,12 @@ public class Artikkeli {
         this.id = id;
     }
 
-    public String getNimi() {
-        return nimi;
-    }
-
-    public void setNimi(String nimi) {
-        this.nimi = nimi;
-    }
-
-    public int getPituus() {
-        return pituus;
-    }
-
-    public void setPituus(int pituus) {
-        this.pituus = pituus;
-    }
-
-    public String getLinkki() {
-        return linkki;
-    }
-
-    public void setLinkki(String linkki) {
-        this.linkki = linkki;
-    }
-
-    public String getTekija() {
-        return tekija;
-    }
-
-    public void setTekija(String tekija) {
-        this.tekija = tekija;
-    }
-
-    public LocalDate getPaivamaara() {
-        return paivamaara;
-    }
-
-    public void setPaivamaara(LocalDate paivamaara) {
-        this.paivamaara = paivamaara;
-    }
-
     public String toString() {
-        String tuloste =  nimi + ", Lehti: " + julkaisuLehti + ", nro: " + numero + ", sivuja: " + sivut + ", tekijä: " + tekija + ", julkaistu: " + julkaisuVuosi;
-        if (!linkki.isEmpty()) {
-            tuloste += ", linkki: " + linkki;
+        String tuloste = super.getNimi() + ", Lehti: " + julkaisuLehti + ", nro: " + numero + ", sivuja: " + sivut + ", tekijä: " + super.getTekija() + ", julkaistu: " + super.getJulkaisuVuosi();
+        if (!super.getLinkki().isEmpty()) {
+            tuloste += ", linkki: " + super.getLinkki();
         }
-        if (luettu) {
+        if (super.isLuettu()) {
             tuloste += ", luettu";
         } else {
             tuloste += ", lukematon";
